EnvelopeSegmentation: refactor into feature curve + detector
